backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
email-validator
python-dotenv
python-jose
passlib==1.7.4
bcrypt==3.2.2
python-multipart
requests
pydantic-settings
yfinance

